Added isStarted to Changelog Changes requested in code review
