+++ Bugsnag/Client/BugsnagClient+Private.h
@@ -41,6 +41,7 @@ BSG_OBJC_DIRECT_MEMBERS
 @property (nonatomic) NSMutableDictionary *extraRuntimeInfo;
 
 @property (atomic) BOOL started;
+@property (atomic) BOOL ready;
 
 /// A read-only version of started property. It was created to match the Bugsnag API
 @property (atomic, readonly) BOOL isStarted;
